voicehive-hotels/services/test-requirements.txt
# Test requirements for VoiceHive Hotels services

# Testing frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# FastAPI testing
httpx==0.25.2
websocket-client==1.6.4

# Mocking
unittest-mock==1.5.0

# For parsing Prometheus metrics in tests
prometheus-client==0.19.0
